requirements.txt
requests
pandas
python-dotenv
openai>=1.0.0
